just some cleanup and simple changes
