+++ org.hl7.fhir.core.generator/configuration/configuration.ini
@@ -30,7 +30,7 @@ Resource = org.hl7.fhir.instance.model.api.IAnyResource
 Period = ca.uhn.fhir.model.api.TemporalPrecisionEnum
 Extension = org.hl7.fhir.instance.model.api.IBaseExtension, org.hl7.fhir.instance.model.api.IBaseDatatype, org.hl7.fhir.instance.model.api.IBaseHasExtensions
 HumanName = ca.uhn.fhir.util.DatatypeUtil, org.hl7.fhir.instance.model.api.IPrimitiveType
-StructureMap = org.hl7.fhir.r5.utils.StructureMapUtilities
+StructureMap = org.hl7.fhir.r5.utils.structuredmap.StructureMapUtilities
 Narrative = org.hl7.fhir.instance.model.api.INarrative
 Coding = org.hl7.fhir.instance.model.api.IBaseCoding
 OperationOutcome = org.hl7.fhir.instance.model.api.IBaseOperationOutcome
