+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/MappingSheetParser.java
@@ -31,9 +31,9 @@ import org.hl7.fhir.r5.model.StructureMap.StructureMapGroupRuleComponent;
 import org.hl7.fhir.r5.model.StructureMap.StructureMapGroupRuleDependentComponent;
 import org.hl7.fhir.r5.model.StructureMap.StructureMapGroupRuleSourceComponent;
 import org.hl7.fhir.r5.model.StructureMap.StructureMapGroupRuleTargetComponent;
-import org.hl7.fhir.r5.model.StructureMap.StructureMapGroupTypeMode;
 import org.hl7.fhir.r5.model.StructureMap.StructureMapTransform;
 import org.hl7.fhir.r5.model.UrlType;
+import org.hl7.fhir.r5.utils.structuredmap.StructureMapUtilities;
 import org.hl7.fhir.utilities.CSVReader;
 import org.hl7.fhir.utilities.TextFile;
 import org.hl7.fhir.utilities.Utilities;
