+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/structuredmap/ResolvedGroup.java
@@ -1,0 +1,8 @@
+package org.hl7.fhir.r5.utils.structuredmap;
+
+import org.hl7.fhir.r5.model.StructureMap;
+
+public class ResolvedGroup {
+  public StructureMap.StructureMapGroupComponent target;
+  public StructureMap targetMap;
+}
