+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/structuredmap/SourceElementComponentWrapper.java
@@ -1,0 +1,22 @@
+package org.hl7.fhir.r5.utils.structuredmap;
+
+import org.hl7.fhir.r5.model.ConceptMap;
+
+public class SourceElementComponentWrapper {
+  private ConceptMap.ConceptMapGroupComponent group;
+  private ConceptMap.SourceElementComponent comp;
+
+  public SourceElementComponentWrapper(ConceptMap.ConceptMapGroupComponent group, ConceptMap.SourceElementComponent comp) {
+    super();
+    this.group = group;
+    this.comp = comp;
+  }
+
+  public ConceptMap.ConceptMapGroupComponent getGroup() {
+    return group;
+  }
+
+  public ConceptMap.SourceElementComponent getComp() {
+    return comp;
+  }
+}
