+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/structuredmap/StringPair.java
@@ -1,0 +1,20 @@
+package org.hl7.fhir.r5.utils.structuredmap;
+
+public class StringPair {
+  private String var;
+  private String desc;
+
+  public StringPair(String var, String desc) {
+    super();
+    this.var = var;
+    this.desc = desc;
+  }
+
+  public String getVar() {
+    return var;
+  }
+
+  public String getDesc() {
+    return desc;
+  }
+}
