+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/structuredmap/TransformContext.java
@@ -1,0 +1,15 @@
+package org.hl7.fhir.r5.utils.structuredmap;
+
+public class TransformContext {
+  private Object appInfo;
+
+  public TransformContext(Object appInfo) {
+    super();
+    this.appInfo = appInfo;
+  }
+
+  public Object getAppInfo() {
+    return appInfo;
+  }
+
+}
