+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/structuredmap/VariableMode.java
@@ -1,0 +1,5 @@
+package org.hl7.fhir.r5.utils.structuredmap;
+
+public enum VariableMode {
+  INPUT, OUTPUT, SHARED
+}
