+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/Content.java
@@ -1,0 +1,8 @@
+package org.hl7.fhir.validation;
+
+import org.hl7.fhir.r5.elementmodel.Manager;
+
+class Content {
+  byte[] focus = null;
+  Manager.FhirFormat cntType = null;
+}
