+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/ValidationEngine.java
@@ -1,55 +1,9 @@
 package org.hl7.fhir.validation;
 
-import java.io.BufferedOutputStream;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.PrintWriter;
-import java.net.HttpURLConnection;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.net.URLConnection;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.UUID;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
-
-import org.apache.commons.io.IOUtils;
-import org.hl7.fhir.convertors.VersionConvertorAdvisor50;
-import org.hl7.fhir.convertors.VersionConvertor_10_30;
-import org.hl7.fhir.convertors.VersionConvertor_10_40;
-import org.hl7.fhir.convertors.VersionConvertor_10_50;
-import org.hl7.fhir.convertors.VersionConvertor_14_30;
-import org.hl7.fhir.convertors.VersionConvertor_14_40;
-import org.hl7.fhir.convertors.VersionConvertor_14_50;
-import org.hl7.fhir.convertors.VersionConvertor_30_40;
-import org.hl7.fhir.convertors.VersionConvertor_30_50;
-import org.hl7.fhir.convertors.VersionConvertor_40_50;
+import com.google.gson.JsonObject;
+import org.hl7.fhir.convertors.*;
 import org.hl7.fhir.convertors.loaders.BaseLoaderR5.NullLoaderKnowledgeProvider;
-import org.hl7.fhir.convertors.loaders.R2016MayToR5Loader;
-import org.hl7.fhir.convertors.loaders.R2ToR5Loader;
-import org.hl7.fhir.convertors.loaders.R3ToR5Loader;
-import org.hl7.fhir.convertors.loaders.R4ToR5Loader;
-import org.hl7.fhir.convertors.loaders.R5ToR5Loader;
+import org.hl7.fhir.convertors.loaders.*;
 import org.hl7.fhir.convertors.txClient.TerminologyClientFactory;
 import org.hl7.fhir.exceptions.DefinitionException;
 import org.hl7.fhir.exceptions.FHIRException;
@@ -60,49 +14,27 @@ import org.hl7.fhir.r5.context.SimpleWorkerContext;
 import org.hl7.fhir.r5.elementmodel.Element;
 import org.hl7.fhir.r5.elementmodel.Manager;
 import org.hl7.fhir.r5.elementmodel.Manager.FhirFormat;
-import org.hl7.fhir.r5.elementmodel.ParserBase.ValidationPolicy;
 import org.hl7.fhir.r5.elementmodel.ObjectConverter;
 import org.hl7.fhir.r5.formats.FormatUtilities;
 import org.hl7.fhir.r5.formats.IParser.OutputStyle;
 import org.hl7.fhir.r5.formats.JsonParser;
 import org.hl7.fhir.r5.formats.XmlParser;
-import org.hl7.fhir.r5.model.Base;
-import org.hl7.fhir.r5.model.Bundle;
-import org.hl7.fhir.r5.model.CanonicalResource;
-import org.hl7.fhir.r5.model.Coding;
-import org.hl7.fhir.r5.model.Constants;
-import org.hl7.fhir.r5.model.DomainResource;
-import org.hl7.fhir.r5.model.FhirPublication;
-import org.hl7.fhir.r5.model.ImplementationGuide;
+import org.hl7.fhir.r5.model.*;
 import org.hl7.fhir.r5.model.ImplementationGuide.ImplementationGuideGlobalComponent;
-import org.hl7.fhir.r5.model.OperationOutcome;
 import org.hl7.fhir.r5.model.OperationOutcome.OperationOutcomeIssueComponent;
-import org.hl7.fhir.r5.model.Parameters;
-import org.hl7.fhir.r5.model.Reference;
-import org.hl7.fhir.r5.model.Resource;
-import org.hl7.fhir.r5.model.StructureDefinition;
-import org.hl7.fhir.r5.model.StructureMap;
 import org.hl7.fhir.r5.renderers.RendererFactory;
 import org.hl7.fhir.r5.renderers.utils.RenderingContext;
 import org.hl7.fhir.r5.renderers.utils.RenderingContext.ResourceRendererMode;
 import org.hl7.fhir.r5.terminologies.ConceptMapEngine;
 import org.hl7.fhir.r5.utils.EOperationOutcome;
 import org.hl7.fhir.r5.utils.FHIRPathEngine;
-import org.hl7.fhir.r5.utils.IResourceValidator.BestPracticeWarningLevel;
-import org.hl7.fhir.r5.utils.IResourceValidator.BundleValidationRule;
-import org.hl7.fhir.r5.utils.IResourceValidator.CheckDisplayOption;
-import org.hl7.fhir.r5.utils.IResourceValidator.IValidatorResourceFetcher;
-import org.hl7.fhir.r5.utils.IResourceValidator.IdStatus;
-import org.hl7.fhir.r5.utils.IResourceValidator.ReferenceValidationPolicy;
+import org.hl7.fhir.r5.utils.IResourceValidator.*;
 import org.hl7.fhir.r5.utils.OperationOutcomeUtilities;
-import org.hl7.fhir.r5.utils.StructureMapUtilities;
-import org.hl7.fhir.r5.utils.StructureMapUtilities.ITransformerServices;
 import org.hl7.fhir.r5.utils.ToolingExtensions;
-import org.hl7.fhir.utilities.IniFile;
-import org.hl7.fhir.utilities.TextFile;
+import org.hl7.fhir.r5.utils.structuredmap.ITransformerServices;
+import org.hl7.fhir.r5.utils.structuredmap.StructureMapUtilities;
 import org.hl7.fhir.utilities.TimeTracker;
-import org.hl7.fhir.utilities.Utilities;
-import org.hl7.fhir.utilities.VersionUtilities;
+import org.hl7.fhir.utilities.*;
 import org.hl7.fhir.utilities.i18n.I18nConstants;
 import org.hl7.fhir.utilities.json.JSONUtil;
 import org.hl7.fhir.utilities.json.JsonTrackingParser;
@@ -111,22 +43,28 @@ import org.hl7.fhir.utilities.npm.NpmPackage;
 import org.hl7.fhir.utilities.npm.ToolsVersion;
 import org.hl7.fhir.utilities.turtle.Turtle;
 import org.hl7.fhir.utilities.validation.ValidationMessage;
-import org.hl7.fhir.utilities.validation.ValidationMessage.IssueSeverity;
-import org.hl7.fhir.utilities.validation.ValidationMessage.IssueType;
-import org.hl7.fhir.utilities.validation.ValidationMessage.Source;
 import org.hl7.fhir.utilities.xhtml.XhtmlComposer;
 import org.hl7.fhir.utilities.xml.XMLUtil;
 import org.hl7.fhir.validation.BaseValidator.ValidationControl;
-import org.hl7.fhir.validation.ValidationEngine.ValidationRecord;
 import org.hl7.fhir.validation.cli.model.ScanOutputItem;
 import org.hl7.fhir.validation.cli.services.StandAloneValidatorFetcher.IPackageInstaller;
 import org.hl7.fhir.validation.cli.utils.*;
 import org.hl7.fhir.validation.instance.InstanceValidator;
 import org.w3c.dom.Document;
 import org.xml.sax.SAXException;
-import org.xmlpull.v1.builder.XmlDocument;
 
-import com.google.gson.JsonObject;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import java.io.*;
+import java.net.HttpURLConnection;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.net.URLConnection;
+import java.util.*;
+import java.util.Map.Entry;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
 
 /*
@@ -160,91 +98,46 @@ POSSIBILITY OF SUCH DAMAGE.
 
 /**
  * This is just a wrapper around the InstanceValidator class for convenient use
- *
+ * <p>
  * The following resource formats are supported: XML, JSON, Turtle
  * The following versions are supported: 1.0.2, 1.4.0, 3.0.2, 4.0.1, and current
- *
+ * <p>
  * Note: the validation engine is intended to be threadsafe
  * To Use:
- *
+ * <p>
  * 1/ Initialize
- *    ValidationEngine validator = new ValidationEngine(src);
- *      - this must be the packageId of the relevant core specification
- *        for the version you want to validate against (e.g. hl7.fhir.r4.core)
- *
- *    validator.connectToTSServer(txServer);
- *      - this is optional; in the absence of a terminology service, snomed, loinc etc will not be validated
- *
- *    validator.loadIg(src);
- *      - call this any number of times for the Implementation Guide(s) of interest.
- *      - See https://confluence.hl7.org/display/FHIR/Using+the+FHIR+Validator for documentation about the src parameter (-ig parameter)
- *
- *    validator.loadQuestionnaire(src)
- *      - url or filename of a questionnaire to load. Any loaded questionnaires will be used while validating
- *
- *    validator.setNative(doNative);
- *      - whether to do xml/json/rdf schema validation as well
- *
- *   You only need to do this initialization once. You can validate as many times as you like
- *
+ * ValidationEngine validator = new ValidationEngine(src);
+ * - this must be the packageId of the relevant core specification
+ * for the version you want to validate against (e.g. hl7.fhir.r4.core)
+ * <p>
+ * validator.connectToTSServer(txServer);
+ * - this is optional; in the absence of a terminology service, snomed, loinc etc will not be validated
+ * <p>
+ * validator.loadIg(src);
+ * - call this any number of times for the Implementation Guide(s) of interest.
+ * - See https://confluence.hl7.org/display/FHIR/Using+the+FHIR+Validator for documentation about the src parameter (-ig parameter)
+ * <p>
+ * validator.loadQuestionnaire(src)
+ * - url or filename of a questionnaire to load. Any loaded questionnaires will be used while validating
+ * <p>
+ * validator.setNative(doNative);
+ * - whether to do xml/json/rdf schema validation as well
+ * <p>
+ * You only need to do this initialization once. You can validate as many times as you like
+ * <p>
  * 2. validate
- *    validator.validate(src, profiles);
- *      - source (as stream, byte[]), or url or filename of a resource to validate.
- *        Also validate against any profiles (as canonical URLS, equivalent to listing them in Resource.meta.profile)
- *
- *        if the source is provided as byte[] or stream, you need to provide a format too, though you can
- *        leave that as null, and the validator will guess
- *
+ * validator.validate(src, profiles);
+ * - source (as stream, byte[]), or url or filename of a resource to validate.
+ * Also validate against any profiles (as canonical URLS, equivalent to listing them in Resource.meta.profile)
+ * <p>
+ * if the source is provided as byte[] or stream, you need to provide a format too, though you can
+ * leave that as null, and the validator will guess
+ * <p>
  * 3. Or, instead of validating, transform (see documentation and use in Validator.java)
  *
  * @author Grahame Grieve
- *
  */
 public class ValidationEngine implements IValidatorResourceFetcher, IPackageInstaller {
-
-  public class ValidationRecord {
-
-    private String location;
-    private List<ValidationMessage> messages;
-    int err = 0;
-    int warn = 0;
-    int info = 0;
-
-    public ValidationRecord(String location, List<ValidationMessage> messages) {
-      this.location = location;
-      this.messages = messages;
-      for (ValidationMessage vm : messages) {
-        if (vm.getLevel().equals(ValidationMessage.IssueSeverity.FATAL)||vm.getLevel().equals(ValidationMessage.IssueSeverity.ERROR))
-          err++;
-        else if (vm.getLevel().equals(ValidationMessage.IssueSeverity.WARNING))
-          warn++;
-        else if (!vm.isSignpost()) {
-          info++;
-        }
-      }
-    }
-
-    public String getLocation() {
-      return location;
-    }
-
-    public List<ValidationMessage> getMessages() {
-      return messages;
-    }
-
-    public int getErr() {
-      return err;
-    }
-
-    public int getWarn() {
-      return warn;
-    }
-
-    public int getInfo() {
-      return info;
-    }
-
-  }
 
   public class TransformSupportServices implements ITransformerServices {
 
@@ -292,7 +185,6 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   }
 
   private SimpleWorkerContext context;
-//  private FHIRPathEngine fpe;
   private Map<String, byte[]> binaries = new HashMap<String, byte[]>();
   private boolean doNative;
   private boolean noInvariantChecks;
@@ -321,34 +213,6 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     initContext(null);
   }
 
-  public String setTerminologyServer(String src, String log, FhirPublication version) throws FHIRException, URISyntaxException {
-    return connectToTSServer(src, log, version);
-  }
-
-  public boolean isHintAboutNonMustSupport() {
-    return hintAboutNonMustSupport;
-  }
-
-  public void setHintAboutNonMustSupport(boolean hintAboutNonMustSupport) {
-    this.hintAboutNonMustSupport = hintAboutNonMustSupport;
-  }
-
-  public boolean isAnyExtensionsAllowed() {
-    return anyExtensionsAllowed;
-  }
-
-  public void setAnyExtensionsAllowed(boolean anyExtensionsAllowed) {
-    this.anyExtensionsAllowed = anyExtensionsAllowed;
-  }
-
-  public boolean isShowTimes() {
-    return showTimes;
-  }
-
-  public void setShowTimes(boolean showTimes) {
-    this.showTimes = showTimes;
-  }
-
   public ValidationEngine(String src, String txsrvr, String txLog, FhirPublication version, boolean canRunWithoutTerminologyServer, String vString) throws FHIRException, IOException, URISyntaxException {
     pcm = new FilesystemPackageCacheManager(true, ToolsVersion.TOOLS_VERSION);
     loadCoreDefinitions(src, false, null);
@@ -374,14 +238,6 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   public ValidationEngine(String src) throws FHIRException, IOException {
     loadCoreDefinitions(src, false, null);
     pcm = new FilesystemPackageCacheManager(true, ToolsVersion.TOOLS_VERSION);
-  }
-
-  public String getLanguage() {
-    return language;
-  }
-
-  public void setLanguage(String language) {
-    this.language = language;
   }
 
   private void loadCoreDefinitions(String src, boolean recursive, TimeTracker tt) throws FHIRException, IOException {
@@ -423,15 +279,15 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     if (Utilities.noString(version))
       return null;
     if (version.startsWith("1.0"))
-      return new R2ToR5Loader(new String[] { "Conformance", "StructureDefinition", "ValueSet", "SearchParameter", "OperationDefinition", "Questionnaire","ConceptMap","StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
+      return new R2ToR5Loader(new String[]{"Conformance", "StructureDefinition", "ValueSet", "SearchParameter", "OperationDefinition", "Questionnaire", "ConceptMap", "StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
     if (version.startsWith("1.4"))
-      return new R2016MayToR5Loader(new String[] { "Conformance", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire","ConceptMap","StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider()); // special case
+      return new R2016MayToR5Loader(new String[]{"Conformance", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire", "ConceptMap", "StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider()); // special case
     if (version.startsWith("3.0"))
-      return new R3ToR5Loader(new String[] { "CapabilityStatement", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire","ConceptMap","StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
+      return new R3ToR5Loader(new String[]{"CapabilityStatement", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire", "ConceptMap", "StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
     if (version.startsWith("4.0"))
-      return new R4ToR5Loader(new String[] { "CapabilityStatement", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire","ConceptMap","StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
+      return new R4ToR5Loader(new String[]{"CapabilityStatement", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire", "ConceptMap", "StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
     if (version.startsWith("5.0"))
-      return new R5ToR5Loader(new String[] { "CapabilityStatement", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire","ConceptMap","StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
+      return new R5ToR5Loader(new String[]{"CapabilityStatement", "StructureDefinition", "ValueSet", "CodeSystem", "SearchParameter", "OperationDefinition", "Questionnaire", "ConceptMap", "StructureMap", "NamingSystem"}, new NullLoaderKnowledgeProvider());
     return null;
   }
 
@@ -452,14 +308,17 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   }
 
   private Parameters makeExpProfile() {
-    Parameters ep  = new Parameters();
+    Parameters ep = new Parameters();
     ep.addParameter("profile-url", "http://hl7.org/fhir/ExpansionProfile/dc8fd4bc-091a-424a-8a3b-6198ef146891"); // change this to blow the cache
     // all defaults....
     return ep;
   }
 
-  /** explore should be true if we're trying to load an -ig parameter, and false if we're loading source
-   * @throws IOException **/
+  /**
+   * explore should be true if we're trying to load an -ig parameter, and false if we're loading source
+   *
+   * @throws IOException
+   **/
 
   private Map<String, byte[]> loadIgSource(String src, boolean recursive, boolean explore) throws FHIRException, IOException {
     // src can be one of the following:
@@ -470,14 +329,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     if (Common.isNetworkPath(src)) {
       String v = null;
       if (src.contains("|")) {
-        v = src.substring(src.indexOf("|")+1);
+        v = src.substring(src.indexOf("|") + 1);
         src = src.substring(0, src.indexOf("|"));
       }
       String pid = explore ? pcm.getPackageId(src) : null;
       if (!Utilities.noString(pid))
-        return fetchByPackage(pid+(v == null ? "" : "#"+v));
+        return fetchByPackage(pid + (v == null ? "" : "#" + v));
       else
-        return fetchFromUrl(src+(v == null ? "" : "|"+v), explore);
+        return fetchFromUrl(src + (v == null ? "" : "|" + v), explore);
     }
 
     File f = new File(Utilities.path(src));
@@ -499,35 +358,35 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       FhirFormat fmt = checkIsResource(src);
       if (fmt != null) {
         Map<String, byte[]> res = new HashMap<String, byte[]>();
-        res.put(Utilities.changeFileExt(src, "."+fmt.getExtension()), TextFile.fileToBytesNCS(src));
+        res.put(Utilities.changeFileExt(src, "." + fmt.getExtension()), TextFile.fileToBytesNCS(src));
         return res;
       }
     } else if ((src.matches(FilesystemPackageCacheManager.PACKAGE_REGEX) || src.matches(FilesystemPackageCacheManager.PACKAGE_VERSION_REGEX)) && !src.endsWith(".zip") && !src.endsWith(".tgz")) {
       return fetchByPackage(src);
     }
-    throw new FHIRException("Unable to find/resolve/read "+(explore ? "-ig " : "")+src);
+    throw new FHIRException("Unable to find/resolve/read " + (explore ? "-ig " : "") + src);
   }
 
   private Map<String, byte[]> loadIgSourceForVersion(String src, boolean recursive, boolean explore, VersionSourceInformation versions) throws FHIRException, IOException {
     if (Common.isNetworkPath(src)) {
       String v = null;
       if (src.contains("|")) {
-        v = src.substring(src.indexOf("|")+1);
+        v = src.substring(src.indexOf("|") + 1);
         src = src.substring(0, src.indexOf("|"));
       }
       String pid = pcm.getPackageId(src);
       if (!Utilities.noString(pid)) {
-        versions.see(fetchVersionByPackage(pid+(v == null ? "" : "#"+v)), "Package "+src);
+        versions.see(fetchVersionByPackage(pid + (v == null ? "" : "#" + v)), "Package " + src);
         return null;
       } else {
-        return fetchVersionFromUrl(src+(v == null ? "" : "|"+v), explore, versions);
+        return fetchVersionFromUrl(src + (v == null ? "" : "|" + v), explore, versions);
       }
     }
 
     File f = new File(Utilities.path(src));
     if (f.exists()) {
       if (f.isDirectory() && new File(Utilities.path(src, "package.tgz")).exists()) {
-        versions.see(loadPackageForVersion(new FileInputStream(Utilities.path(src, "package.tgz")), Utilities.path(src, "package.tgz")), "Package "+src);
+        versions.see(loadPackageForVersion(new FileInputStream(Utilities.path(src, "package.tgz")), Utilities.path(src, "package.tgz")), "Package " + src);
         return null;
       }
       if (f.isDirectory() && new File(Utilities.path(src, "igpack.zip")).exists())
@@ -537,7 +396,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       if (f.isDirectory())
         return scanDirectory(f, recursive);
       if (src.endsWith(".tgz")) {
-        versions.see(loadPackageForVersion(new FileInputStream(src), src), "Package "+src);
+        versions.see(loadPackageForVersion(new FileInputStream(src), src), "Package " + src);
         return null;
       }
       if (src.endsWith(".pack"))
@@ -547,14 +406,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       FhirFormat fmt = checkIsResource(src);
       if (fmt != null) {
         Map<String, byte[]> res = new HashMap<String, byte[]>();
-        res.put(Utilities.changeFileExt(src, "."+fmt.getExtension()), TextFile.fileToBytesNCS(src));
+        res.put(Utilities.changeFileExt(src, "." + fmt.getExtension()), TextFile.fileToBytesNCS(src));
         return res;
       }
     } else if ((src.matches(FilesystemPackageCacheManager.PACKAGE_REGEX) || src.matches(FilesystemPackageCacheManager.PACKAGE_VERSION_REGEX)) && !src.endsWith(".zip") && !src.endsWith(".tgz")) {
-      versions.see(fetchVersionByPackage(src), "Package "+src);
+      versions.see(fetchVersionByPackage(src), "Package " + src);
       return null;
     }
-    throw new FHIRException("Unable to find/resolve/read -ig "+src);
+    throw new FHIRException("Unable to find/resolve/read -ig " + src);
   }
 
 
@@ -594,21 +453,21 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       }
     }
     if (cnt == null) {
-      throw new FHIRException("Unable to fetch content from "+src+" ("+errors.toString()+")");
+      throw new FHIRException("Unable to fetch content from " + src + " (" + errors.toString() + ")");
 
     }
     FhirFormat fmt = checkFormat(cnt, src);
     if (fmt != null) {
       Map<String, byte[]> res = new HashMap<String, byte[]>();
-      res.put(Utilities.changeFileExt(src, "."+fmt.getExtension()), cnt);
+      res.put(Utilities.changeFileExt(src, "." + fmt.getExtension()), cnt);
       return res;
     }
-    throw new FHIRException("Unable to read content from "+src+": cannot determine format");
+    throw new FHIRException("Unable to read content from " + src + ": cannot determine format");
   }
 
   private Map<String, byte[]> fetchVersionFromUrl(String src, boolean explore, VersionSourceInformation versions) throws FHIRException, IOException {
     if (src.endsWith(".tgz")) {
-      versions.see(loadPackageForVersion(fetchFromUrlSpecific(src, false), src), "From Package "+src);
+      versions.see(loadPackageForVersion(fetchFromUrlSpecific(src, false), src), "From Package " + src);
       return null;
     }
     if (src.endsWith(".pack"))
@@ -620,7 +479,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     if (explore) {
       stream = fetchFromUrlSpecific(Utilities.pathURL(src, "package.tgz"), true);
       if (stream != null) {
-        versions.see(loadPackageForVersion(stream, Utilities.pathURL(src, "package.tgz")), "From Package at "+src);
+        versions.see(loadPackageForVersion(stream, Utilities.pathURL(src, "package.tgz")), "From Package at " + src);
         return null;
       }
       // todo: these options are deprecated - remove once all IGs have been rebuilt post R4 technical correction
@@ -644,15 +503,15 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     FhirFormat fmt = checkIsResource(cnt, src);
     if (fmt != null) {
       Map<String, byte[]> res = new HashMap<String, byte[]>();
-      res.put(Utilities.changeFileExt(src, "."+fmt.getExtension()), cnt);
+      res.put(Utilities.changeFileExt(src, "." + fmt.getExtension()), cnt);
       return res;
     }
     String fn = Utilities.path("[tmp]", "fetch-resource-error-content.bin");
     TextFile.bytesToFile(cnt, fn);
-    System.out.println("Error Fetching "+src);
-    System.out.println("Some content was found, saved to "+fn);
-    System.out.println("1st 100 bytes = "+presentForDebugging(cnt));
-    throw new FHIRException("Unable to find/resolve/read "+(explore ? "-ig " : "")+src);
+    System.out.println("Error Fetching " + src);
+    System.out.println("Some content was found, saved to " + fn);
+    System.out.println("1st 100 bytes = " + presentForDebugging(cnt));
+    throw new FHIRException("Unable to find/resolve/read " + (explore ? "-ig " : "") + src);
   }
 
   private String presentForDebugging(byte[] cnt) {
@@ -665,7 +524,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   private InputStream fetchFromUrlSpecific(String source, boolean optional) throws FHIRException, IOException {
     try {
-      URL url = new URL(source+"?nocache=" + System.currentTimeMillis());
+      URL url = new URL(source + "?nocache=" + System.currentTimeMillis());
       URLConnection c = url.openConnection();
       return c.getInputStream();
     } catch (IOException e) {
@@ -680,7 +539,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     try {
       try {
         // try with cache-busting option and then try withhout in case the server doesn't support that
-        URL url = new URL(source+"?nocache=" + System.currentTimeMillis());
+        URL url = new URL(source + "?nocache=" + System.currentTimeMillis());
         HttpURLConnection conn = (HttpURLConnection) url.openConnection();
         conn.setRequestProperty("Accept", contentType);
         return TextFile.streamToBytes(conn.getInputStream());
@@ -692,7 +551,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       }
     } catch (IOException e) {
       if (errors != null) {
-        errors.add("Error accessing "+source+": "+e.getMessage());
+        errors.add("Error accessing " + source + ": " + e.getMessage());
       }
       if (optional)
         return null;
@@ -704,13 +563,12 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   private Map<String, byte[]> scanDirectory(File f, boolean recursive) throws FileNotFoundException, IOException {
     Map<String, byte[]> res = new HashMap<>();
     for (File ff : f.listFiles()) {
-      if (ff.isDirectory() && recursive){
-          res.putAll(scanDirectory(ff, true));
-      }
-      else if (!isIgnoreFile(ff)) {
+      if (ff.isDirectory() && recursive) {
+        res.putAll(scanDirectory(ff, true));
+      } else if (!isIgnoreFile(ff)) {
         FhirFormat fmt = checkIsResource(ff.getAbsolutePath());
         if (fmt != null) {
-          res.put(Utilities.changeFileExt(ff.getName(), "."+fmt.getExtension()), TextFile.fileToBytes(ff.getAbsolutePath()));
+          res.put(Utilities.changeFileExt(ff.getName(), "." + fmt.getExtension()), TextFile.fileToBytes(ff.getAbsolutePath()));
         }
       }
     }
@@ -718,7 +576,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   }
 
   private boolean isIgnoreFile(File ff) {
-    if (ff.getName().startsWith(".")|| ff.getAbsolutePath().contains(".git")){
+    if (ff.getName().startsWith(".") || ff.getAbsolutePath().contains(".git")) {
       return true;
     }
     return Utilities.existsInList(Utilities.getFileExtension(ff.getName()).toLowerCase(), "md", "css", "js", "png", "gif", "jpg", "html", "tgz", "pack", "zip");
@@ -733,33 +591,33 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   }
 
   public Map<String, byte[]> loadPackage(NpmPackage pi) throws FHIRException, IOException {
-    context.getLoadedPackages().add(pi.name()+"#"+pi.version());
+    context.getLoadedPackages().add(pi.name() + "#" + pi.version());
     Map<String, byte[]> res = new HashMap<String, byte[]>();
     for (String s : pi.dependencies()) {
-      if (s.endsWith(".x") && s.length()>2) {
-        String packageMajorMinor = s.substring(0, s.length()-2);
+      if (s.endsWith(".x") && s.length() > 2) {
+        String packageMajorMinor = s.substring(0, s.length() - 2);
         boolean found = false;
-        for (int i=0; i<context.getLoadedPackages().size() && !found; ++i) {
+        for (int i = 0; i < context.getLoadedPackages().size() && !found; ++i) {
           String loadedPackage = context.getLoadedPackages().get(i);
           if (loadedPackage.startsWith(packageMajorMinor)) {
             found = true;
           }
         }
         if (found)
-          continue ;
+          continue;
       }
-      if (! context.getLoadedPackages().contains(s)) {
+      if (!context.getLoadedPackages().contains(s)) {
         if (!VersionUtilities.isCorePackage(s)) {
-          System.out.println("+  .. load IG from "+s);
+          System.out.println("+  .. load IG from " + s);
           res.putAll(fetchByPackage(s));
         }
       }
     }
 
     for (String s : pi.listResources("CodeSystem", "ConceptMap", "ImplementationGuide", "CapabilityStatement", "SearchParameter", "Conformance", "StructureMap", "ValueSet", "StructureDefinition")) {
-       res.put(s, TextFile.streamToBytes(pi.load("package", s)));
+      res.put(s, TextFile.streamToBytes(pi.load("package", s)));
     }
-    String ini = "[FHIR]\r\nversion="+pi.fhirVersion()+"\r\n";
+    String ini = "[FHIR]\r\nversion=" + pi.fhirVersion() + "\r\n";
     res.put("version.info", ini.getBytes());
     return res;
   }
@@ -769,14 +627,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     ZipInputStream zip = new ZipInputStream(stream);
     ZipEntry ze;
     while ((ze = zip.getNextEntry()) != null) {
-        String name = ze.getName();
-        InputStream in = zip;
-        ByteArrayOutputStream b = new ByteArrayOutputStream();
-        int n;
-        byte[] buf = new byte[1024];
-        while ((n = in.read(buf, 0, 1024)) > -1) {
-          b.write(buf, 0, n);
-        }
+      String name = ze.getName();
+      InputStream in = zip;
+      ByteArrayOutputStream b = new ByteArrayOutputStream();
+      int n;
+      byte[] buf = new byte[1024];
+      while ((n = in.read(buf, 0, 1024)) > -1) {
+        b.write(buf, 0, n);
+      }
       res.put(name, b.toByteArray());
       zip.closeEntry();
     }
@@ -793,7 +651,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     String version = null;
     if (src.contains("#")) {
       id = src.substring(0, src.indexOf("#"));
-      version = src.substring(src.indexOf("#")+1);
+      version = src.substring(src.indexOf("#") + 1);
     }
     if (pcm == null) {
       log("Creating Package manager?");
@@ -806,7 +664,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     if (version == null) {
       pi = pcm.loadPackageFromCacheOnly(id);
       if (pi != null)
-        log("   ... Using version "+pi.version());
+        log("   ... Using version " + pi.version());
     } else
       pi = pcm.loadPackageFromCacheOnly(id, version);
     if (pi == null) {
@@ -820,7 +678,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     String version = null;
     if (src.contains("#")) {
       id = src.substring(0, src.indexOf("#"));
-      version = src.substring(src.indexOf("#")+1);
+      version = src.substring(src.indexOf("#") + 1);
     }
     if (pcm == null) {
       log("Creating Package manager?");
@@ -833,7 +691,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     if (version == null) {
       pi = pcm.loadPackageFromCacheOnly(id);
       if (pi != null)
-        log("   ... Using version "+pi.version());
+        log("   ... Using version " + pi.version());
     } else
       pi = pcm.loadPackageFromCacheOnly(id, version);
     if (pi == null) {
@@ -846,7 +704,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   private Map<String, byte[]> resolvePackage(String id, String v) throws FHIRException, IOException {
     NpmPackage pi = pcm.loadPackage(id, v);
     if (pi != null && v == null)
-      log("   ... Using version "+pi.version());
+      log("   ... Using version " + pi.version());
     return loadPackage(pi);
   }
 
@@ -859,23 +717,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     return context;
   }
 
-
-  public boolean isNoInvariantChecks() {
-    return noInvariantChecks;
-  }
-
-  public void setNoInvariantChecks(boolean value) {
-    this.noInvariantChecks = value;
-  }
-
   private FhirFormat checkFormat(byte[] cnt, String filename) {
-    System.out.println("   ..Detect format for "+filename);
+    System.out.println("   ..Detect format for " + filename);
     try {
       JsonTrackingParser.parseJson(cnt);
       return FhirFormat.JSON;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not JSON: "+e.getMessage());
+        System.out.println("Not JSON: " + e.getMessage());
       }
     }
     try {
@@ -883,7 +732,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       return FhirFormat.XML;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not XML: "+e.getMessage());
+        System.out.println("Not XML: " + e.getMessage());
       }
     }
     try {
@@ -891,7 +740,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       return FhirFormat.TURTLE;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not Turtle: "+e.getMessage());
+        System.out.println("Not Turtle: " + e.getMessage());
       }
     }
     try {
@@ -899,23 +748,23 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       return FhirFormat.TEXT;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not Text: "+e.getMessage());
+        System.out.println("Not Text: " + e.getMessage());
       }
     }
     if (debug)
-      System.out.println("     .. not a resource: "+filename);
+      System.out.println("     .. not a resource: " + filename);
     return null;
   }
 
 
   private FhirFormat checkIsResource(byte[] cnt, String filename) {
-    System.out.println("   ..Detect format for "+filename);
+    System.out.println("   ..Detect format for " + filename);
     try {
       Manager.parse(context, new ByteArrayInputStream(cnt), FhirFormat.JSON);
       return FhirFormat.JSON;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not JSON: "+e.getMessage());
+        System.out.println("Not JSON: " + e.getMessage());
       }
     }
     try {
@@ -923,15 +772,15 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       return FhirFormat.XML;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not XML: "+e.getMessage());
+        System.out.println("Not XML: " + e.getMessage());
       }
     }
     try {
-      Manager.parse(context, new ByteArrayInputStream(cnt),FhirFormat.TURTLE);
+      Manager.parse(context, new ByteArrayInputStream(cnt), FhirFormat.TURTLE);
       return FhirFormat.TURTLE;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not Turtle: "+e.getMessage());
+        System.out.println("Not Turtle: " + e.getMessage());
       }
     }
     try {
@@ -939,11 +788,11 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       return FhirFormat.TEXT;
     } catch (Exception e) {
       if (debug) {
-        System.out.println("Not Text: "+e.getMessage());
+        System.out.println("Not Text: " + e.getMessage());
       }
     }
     if (debug)
-      System.out.println("     .. not a resource: "+filename);
+      System.out.println("     .. not a resource: " + filename);
     return null;
   }
 
@@ -987,7 +836,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         return context.connectToTSServer(TerminologyClientFactory.makeClient(url, version), log);
       } catch (Exception e) {
         if (context.isCanRunWithoutTerminology()) {
-          return "n/a: Running without Terminology Server (error: "+e.getMessage()+")";
+          return "n/a: Running without Terminology Server (error: " + e.getMessage() + ")";
         } else
           throw e;
       }
@@ -1009,7 +858,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   public void scanForIgVersion(String src, boolean recursive, VersionSourceInformation versions) throws IOException, FHIRException, Exception {
     Map<String, byte[]> source = loadIgSourceForVersion(src, recursive, true, versions);
     if (source != null && source.containsKey("version.info"))
-      versions.see(readInfoVersion(source.get("version.info")), "version.info in "+src);
+      versions.see(readInfoVersion(source.get("version.info")), "version.info in " + src);
   }
 
   public void loadIg(String src, boolean recursive) throws IOException, FHIRException {
@@ -1024,10 +873,10 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       }
       System.out.print("  Load " + src);
       if (!src.contains("#")) {
-        System.out.print("#"+npm.version());
+        System.out.print("#" + npm.version());
       }
       int count = context.loadFromPackage(npm, loaderForVersion(npm.fhirVersion()));
-      System.out.println(" - "+count+" resources ("+context.clock().milestone()+")");
+      System.out.println(" - " + count + " resources (" + context.clock().milestone() + ")");
     } else {
       System.out.print("  Load " + src);
       String canonical = null;
@@ -1062,13 +911,13 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       if (canonical != null) {
         grabNatives(source, canonical);
       }
-      System.out.println(" - "+count+" resources ("+context.clock().milestone()+")");
+      System.out.println(" - " + count + " resources (" + context.clock().milestone() + ")");
     }
   }
 
   public Resource loadFileWithErrorChecking(String version, Entry<String, byte[]> t, String fn) {
     if (debug)
-      System.out.print("* load file: "+fn);
+      System.out.print("* load file: " + fn);
     Resource r = null;
     try {
       r = loadResourceByVersion(version, t.getValue(), fn);
@@ -1076,9 +925,9 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         System.out.println(" .. success");
     } catch (Exception e) {
       if (!debug) {
-        System.out.print("* load file: "+fn);
+        System.out.print("* load file: " + fn);
       }
-      System.out.println(" - ignored due to error: "+(e.getMessage() == null ? " (null - NPE)" :  e.getMessage()));
+      System.out.println(" - ignored due to error: " + (e.getMessage() == null ? " (null - NPE)" : e.getMessage()));
       if (debug || ((e.getMessage() != null && e.getMessage().contains("cannot be cast")))) {
         e.printStackTrace();
       }
@@ -1094,10 +943,10 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         res = new org.hl7.fhir.dstu3.formats.XmlParser().parse(new ByteArrayInputStream(content));
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         res = new org.hl7.fhir.dstu3.formats.JsonParser().parse(new ByteArrayInputStream(content));
-      else if (fn.endsWith(".txt") || fn.endsWith(".map") )
+      else if (fn.endsWith(".txt") || fn.endsWith(".map"))
         res = new org.hl7.fhir.dstu3.utils.StructureMapUtilities(null).parse(new String(content));
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
       r = VersionConvertor_30_50.convertResource(res, false);
     } else if (version.startsWith("4.0")) {
       org.hl7.fhir.r4.model.Resource res;
@@ -1105,10 +954,10 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         res = new org.hl7.fhir.r4.formats.XmlParser().parse(new ByteArrayInputStream(content));
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         res = new org.hl7.fhir.r4.formats.JsonParser().parse(new ByteArrayInputStream(content));
-      else if (fn.endsWith(".txt") || fn.endsWith(".map") )
+      else if (fn.endsWith(".txt") || fn.endsWith(".map"))
         res = new org.hl7.fhir.r4.utils.StructureMapUtilities(null).parse(new String(content), fn);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
       r = VersionConvertor_40_50.convertResource(res);
     } else if (version.startsWith("1.4")) {
       org.hl7.fhir.dstu2016may.model.Resource res;
@@ -1117,7 +966,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         res = new org.hl7.fhir.dstu2016may.formats.JsonParser().parse(new ByteArrayInputStream(content));
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
       r = VersionConvertor_14_50.convertResource(res);
     } else if (version.startsWith("1.0")) {
       org.hl7.fhir.dstu2.model.Resource res;
@@ -1126,7 +975,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         res = new org.hl7.fhir.dstu2.formats.JsonParser().parse(new ByteArrayInputStream(content));
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
       VersionConvertorAdvisor50 advisor = new org.hl7.fhir.convertors.misc.IGR2ConvertorAdvisor5();
       r = VersionConvertor_10_50.convertResource(res, advisor);
     } else if (version.equals(Constants.VERSION) || "current".equals(version)) {
@@ -1136,12 +985,12 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         r = new JsonParser().parse(new ByteArrayInputStream(content));
       else if (fn.endsWith(".txt"))
         r = new StructureMapUtilities(context, null, null).parse(TextFile.bytesToString(content), fn);
-      else if (fn.endsWith(".txt") || fn.endsWith(".map") )
-        r = new org.hl7.fhir.r5.utils.StructureMapUtilities(null).parse(new String(content), fn);
+      else if (fn.endsWith(".txt") || fn.endsWith(".map"))
+        r = new StructureMapUtilities(null).parse(new String(content), fn);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
     } else
-      throw new FHIRException("Unsupported version "+version);
+      throw new FHIRException("Unsupported version " + version);
     return r;
   }
 
@@ -1159,29 +1008,16 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   private void grabNatives(Map<String, byte[]> source, String prefix) {
     for (Entry<String, byte[]> e : source.entrySet()) {
       if (e.getKey().endsWith(".zip"))
-        binaries.put(prefix+"#"+e.getKey(), e.getValue());
+        binaries.put(prefix + "#" + e.getKey(), e.getValue());
     }
-  }
-
-  public void setQuestionnaireMode(QuestionnaireMode questionnaireMode) {
-    this.questionnaireMode = questionnaireMode;
-  }
-
-  public void setNative(boolean doNative) {
-    this.doNative = doNative;
-  }
-
-  private class Content {
-    byte[] focus = null;
-    FhirFormat cntType = null;
   }
 
   public Content loadContent(String source, String opName, boolean asIg) throws FHIRException, IOException {
     Map<String, byte[]> s = loadIgSource(source, false, asIg);
     Content res = new Content();
     if (s.size() != 1)
-      throw new FHIRException("Unable to find resource " + source + " to "+opName);
-    for (Entry<String, byte[]> t: s.entrySet()) {
+      throw new FHIRException("Unable to find resource " + source + " to " + opName);
+    for (Entry<String, byte[]> t : s.entrySet()) {
       res.focus = t.getValue();
       if (t.getKey().endsWith(".json"))
         res.cntType = FhirFormat.JSON;
@@ -1211,7 +1047,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       for (String p : profiles) {
         StructureDefinition sd = context.fetchResource(StructureDefinition.class, p);
         if (sd == null) {
-          throw new Error("Unable to resolve profile "+p);
+          throw new Error("Unable to resolve profile " + p);
         }
         list.add(sd);
       }
@@ -1222,7 +1058,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   public OperationOutcome validate(String source, List<String> profiles) throws FHIRException, IOException {
     List<String> l = new ArrayList<String>();
     l.add(source);
-    return (OperationOutcome)validate(l, profiles, null);
+    return (OperationOutcome) validate(l, profiles, null);
   }
 
   public List<ScanOutputItem> validateScan(List<String> sources, Set<String> guides) throws FHIRException, IOException, EOperationOutcome {
@@ -1237,7 +1073,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       List<ValidationMessage> messages = new ArrayList<ValidationMessage>();
       Element e = null;
       try {
-        System.out.println("Validate "+ref);
+        System.out.println("Validate " + ref);
         messages.clear();
         e = validator.validate(null, messages, new ByteArrayInputStream(cnt.focus), cnt.cntType);
         res.add(new ScanOutputItem(ref, null, null, messagesToOutcome(messages)));
@@ -1248,12 +1084,12 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         String rt = e.fhirType();
         for (String u : guides) {
           ImplementationGuide ig = context.fetchResource(ImplementationGuide.class, u);
-          System.out.println("Check Guide "+ig.getUrl());
+          System.out.println("Check Guide " + ig.getUrl());
           String canonical = ig.getUrl().contains("/Impl") ? ig.getUrl().substring(0, ig.getUrl().indexOf("/Impl")) : ig.getUrl();
           String url = getGlobal(ig, rt);
           if (url != null) {
             try {
-              System.out.println("Validate "+ref+" against "+ig.getUrl());
+              System.out.println("Validate " + ref + " against " + ig.getUrl());
               messages.clear();
               validator.validate(null, messages, new ByteArrayInputStream(cnt.focus), cnt.cntType, url);
               res.add(new ScanOutputItem(ref, ig, null, messagesToOutcome(messages)));
@@ -1267,7 +1103,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
               done.add(sd.getUrl());
               if (sd.getUrl().startsWith(canonical) && rt.equals(sd.getType())) {
                 try {
-                  System.out.println("Validate "+ref+" against "+sd.getUrl());
+                  System.out.println("Validate " + ref + " against " + sd.getUrl());
                   messages.clear();
                   validator.validate(null, messages, new ByteArrayInputStream(cnt.focus), cnt.cntType, asSdList(sd));
                   res.add(new ScanOutputItem(ref, ig, sd, messagesToOutcome(messages)));
@@ -1304,28 +1140,28 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       Content cnt = loadContent(ref, "validate", false);
       String s = TextFile.bytesToString(cnt.focus);
       if (s.contains("http://hl7.org/fhir/3.0")) {
-        versions.see("3.0", "Profile in "+ref);
+        versions.see("3.0", "Profile in " + ref);
       }
       if (s.contains("http://hl7.org/fhir/1.0")) {
-        versions.see("1.0", "Profile in "+ref);
+        versions.see("1.0", "Profile in " + ref);
       }
       if (s.contains("http://hl7.org/fhir/4.0")) {
-        versions.see("4.0", "Profile in "+ref);
+        versions.see("4.0", "Profile in " + ref);
       }
       if (s.contains("http://hl7.org/fhir/1.4")) {
-        versions.see("1.4", "Profile in "+ref);
+        versions.see("1.4", "Profile in " + ref);
       }
       try {
         if (s.startsWith("{")) {
-          JsonObject json = JsonTrackingParser.parse(s,  null);
+          JsonObject json = JsonTrackingParser.parse(s, null);
           if (json.has("fhirVersion")) {
-            versions.see(VersionUtilities.getMajMin(JSONUtil.str(json, "fhirVersion")), "fhirVersion in "+ref);
+            versions.see(VersionUtilities.getMajMin(JSONUtil.str(json, "fhirVersion")), "fhirVersion in " + ref);
           }
         } else {
-          Document doc  = parseXml(cnt.focus);
+          Document doc = parseXml(cnt.focus);
           String v = XMLUtil.getNamedChildValue(doc.getDocumentElement(), "fhirVersion");
           if (v != null) {
-            versions.see(VersionUtilities.getMajMin(v), "fhirVersion in "+ref);
+            versions.see(VersionUtilities.getMajMin(v), "fhirVersion in " + ref);
           }
         }
       } catch (Exception e) {
@@ -1336,7 +1172,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   public Resource validate(List<String> sources, List<String> profiles, List<ValidationRecord> record) throws FHIRException, IOException {
     if (profiles.size() > 0) {
-      System.out.println("  Profiles: "+profiles);
+      System.out.println("  Profiles: " + profiles);
     }
     List<String> refs = new ArrayList<String>();
     boolean asBundle = parseSources(sources, refs);
@@ -1354,7 +1190,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         results.addEntry().setResource(outcome);
         tts.end();
       } catch (Exception e) {
-        System.out.println("Validation Infrastructure fail validating "+ref+": "+e.getMessage());
+        System.out.println("Validation Infrastructure fail validating " + ref + ": " + e.getMessage());
         tts.end();
         throw new FHIRException(e);
       }
@@ -1367,6 +1203,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   /**
    * Iterates through the list of passed in sources, extracting all references and populated them in the passed in list.
+   *
    * @return {@link Boolean#TRUE} if more than one reference is found.
    */
   public boolean parseSources(List<String> sources, List<String> refs) throws IOException {
@@ -1379,6 +1216,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   /**
    * Parses passed in resource path, adding any found references to the passed in list.
+   *
    * @return {@link Boolean#TRUE} if more than one reference is found.
    */
   private boolean extractReferences(String name, List<String> refs) throws IOException {
@@ -1404,14 +1242,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       if (file.isFile()) {
         refs.add(name);
       } else {
-        for (int i=0; i < file.listFiles().length; i++) {
+        for (int i = 0; i < file.listFiles().length; i++) {
           File[] fileList = file.listFiles();
           if (fileList[i].isFile())
             refs.add(fileList[i].getPath());
         }
       }
     }
-    return refs.size() > 1 ;
+    return refs.size() > 1;
   }
 
   public OperationOutcome validate(byte[] source, FhirFormat cntType, List<String> profiles, List<ValidationMessage> messages) throws FHIRException, IOException, EOperationOutcome {
@@ -1430,7 +1268,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     InstanceValidator validator = getValidator();
     validator.validate(null, messages, new ByteArrayInputStream(source), cntType, asSdList(profiles));
     if (showTimes) {
-      System.out.println(location+": "+validator.reportTimes());
+      System.out.println(location + ": " + validator.reportTimes());
     }
     if (record != null) {
       record.add(new ValidationRecord(location, messages));
@@ -1476,7 +1314,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       try {
         fpe.parse(vm.getLocation());
       } catch (Exception e) {
-        System.out.println("Internal error in location for message: '"+e.getMessage()+"', loc = '"+vm.getLocation()+"', err = '"+vm.getMessage()+"'");
+        System.out.println("Internal error in location for message: '" + e.getMessage() + "', loc = '" + vm.getLocation() + "', err = '" + vm.getMessage() + "'");
       }
       op.getIssue().add(OperationOutcomeUtilities.convertToIssue(vm, op));
     }
@@ -1500,7 +1338,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     org.hl7.fhir.r5.elementmodel.Element src = Manager.parse(context, new ByteArrayInputStream(source), cntType);
     StructureMap map = context.getTransform(mapUri);
     if (map == null)
-      throw new Error("Unable to find map "+mapUri+" (Known Maps = "+context.listMapUrls()+")");
+      throw new Error("Unable to find map " + mapUri + " (Known Maps = " + context.listMapUrls() + ")");
 
     org.hl7.fhir.r5.elementmodel.Element resource = getTargetResourceFromStructureMap(map);
     scu.transform(null, src, map, resource);
@@ -1509,26 +1347,26 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   private org.hl7.fhir.r5.elementmodel.Element getTargetResourceFromStructureMap(StructureMap map) {
     String targetTypeUrl = null;
-    for(StructureMap.StructureMapStructureComponent component: map.getStructure()) {
-      if(component.getMode() == StructureMap.StructureMapModelMode.TARGET) {
+    for (StructureMap.StructureMapStructureComponent component : map.getStructure()) {
+      if (component.getMode() == StructureMap.StructureMapModelMode.TARGET) {
         targetTypeUrl = component.getUrl();
         break;
       }
     }
 
-    if(targetTypeUrl == null)
+    if (targetTypeUrl == null)
       throw new FHIRException("Unable to determine resource URL for target type");
 
     StructureDefinition structureDefinition = null;
-    for(StructureDefinition sd:this.context.getStructures()) {
-      if(sd.getUrl().equalsIgnoreCase(targetTypeUrl)) {
+    for (StructureDefinition sd : this.context.getStructures()) {
+      if (sd.getUrl().equalsIgnoreCase(targetTypeUrl)) {
         structureDefinition = sd;
         break;
       }
     }
 
-    if(structureDefinition == null)
-      throw new FHIRException("Unable to find StructureDefinition for target type ('"+targetTypeUrl+"')");
+    if (structureDefinition == null)
+      throw new FHIRException("Unable to find StructureDefinition for target type ('" + targetTypeUrl + "')");
 
     return Manager.build(getContext(), structureDefinition);
   }
@@ -1576,14 +1414,6 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   }
 
-  public String getVersion() {
-    return version;
-  }
-
-  public void setVersion(String version) {
-    this.version = version;
-  }
-
   public InstanceValidator getValidator() {
     InstanceValidator validator = new InstanceValidator(context, null, null);
     validator.setHintAboutNonMustSupport(hintAboutNonMustSupport);
@@ -1598,15 +1428,9 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     validator.setFetcher(this);
     validator.getImplementationGuides().addAll(igs);
     validator.getBundleValidationRules().addAll(bundleValidationRules);
-    validator.getValidationControl().putAll(validationControl );
+    validator.getValidationControl().putAll(validationControl);
     validator.setQuestionnaireMode(questionnaireMode);
     return validator;
-  }
-
-  public void setMapLog(String mapLog) throws FileNotFoundException {
-    if (mapLog != null) {
-      this.mapLog = new PrintWriter(mapLog);
-    }
   }
 
   public void prepare() {
@@ -1614,7 +1438,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       try {
         makeSnapshot(sd);
       } catch (Exception e) {
-        System.out.println("Process Note: Unable to generate snapshot for "+sd.present()+": "+e.getMessage());
+        System.out.println("Process Note: Unable to generate snapshot for " + sd.present() + ": " + e.getMessage());
         if (debug) {
           e.printStackTrace();
         }
@@ -1633,22 +1457,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 
   }
 
-  public boolean isDebug() {
-    return debug;
-  }
-
-  public void setDebug(boolean debug) {
-    this.debug = debug;
-  }
-
   public void genScanOutput(String folder, List<ScanOutputItem> items) throws IOException, FHIRException, EOperationOutcome {
     String f = Utilities.path(folder, "comparison.zip");
     download("http://fhir.org/archive/comparison.zip", f);
     unzip(f, folder);
 
     for (int i = 0; i < items.size(); i++) {
-      items.get(i).setId("c"+Integer.toString(i));
-      genScanOutputItem(items.get(i), Utilities.path(folder, items.get(i).getId()+".html"));
+      items.get(i).setId("c" + Integer.toString(i));
+      genScanOutputItem(items.get(i), Utilities.path(folder, items.get(i).getId() + ".html"));
     }
 
     StringBuilder b = new StringBuilder();
@@ -1697,7 +1513,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     b.append("<tr><th></th><th></th>");
     for (String s : sorted(igs)) {
       ImplementationGuide ig = context.fetchResource(ImplementationGuide.class, s);
-      b.append("<th colspan=\""+Integer.toString(profiles.get(s).size()+1)+"\"><b title=\""+s+"\">"+ig.present()+"</b></th>");
+      b.append("<th colspan=\"" + Integer.toString(profiles.get(s).size() + 1) + "\"><b title=\"" + s + "\">" + ig.present() + "</b></th>");
     }
     b.append("</tr>\r\n");
     b.append("<tr><th><b>Source</b></th><th><span>Core Spec</span></th>");
@@ -1706,14 +1522,14 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       b.append("<th><span>Global</span></th>");
       for (String sp : sorted(profiles.get(s))) {
         StructureDefinition sd = context.fetchResource(StructureDefinition.class, sp);
-        b.append("<th><b title=\""+sp+"\"><span>"+sd.present()+"</span></b></th>");
+        b.append("<th><b title=\"" + sp + "\"><span>" + sd.present() + "</span></b></th>");
       }
     }
     b.append("</tr>\r\n");
 
     for (String s : sorted(refs)) {
       b.append("<tr>");
-      b.append("<td>"+s+"</td>");
+      b.append("<td>" + s + "</td>");
       b.append(genOutcome(items, s, null, null));
       for (String si : sorted(igs)) {
         ImplementationGuide ig = context.fetchResource(ImplementationGuide.class, si);
@@ -1730,7 +1546,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     b.append("<table class=\"grid\">");
     b.append("<tr><th></th><th></th>");
     for (String s : sorted(refs)) {
-      b.append("<th><span>"+s+"</span></th>");
+      b.append("<th><span>" + s + "</span></th>");
     }
     b.append("</tr>\r\n");
     b.append("<tr><td></td><td>Core Spec</td>");
@@ -1741,7 +1557,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     for (String si : sorted(igs)) {
       b.append("<tr>");
       ImplementationGuide ig = context.fetchResource(ImplementationGuide.class, si);
-      b.append("<td><b title=\""+si+"\">"+ig.present()+"</b></td>");
+      b.append("<td><b title=\"" + si + "\">" + ig.present() + "</b></td>");
       b.append("<td>Global</td>");
       for (String s : sorted(refs)) {
         b.append(genOutcome(items, s, si, null));
@@ -1751,7 +1567,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       for (String sp : sorted(profiles.get(ig.getUrl()))) {
         b.append("<tr>");
         StructureDefinition sd = context.fetchResource(StructureDefinition.class, sp);
-        b.append("<td></td><td><b title=\""+sp+"\">"+sd.present()+"</b></td>");
+        b.append("<td></td><td><b title=\"" + sp + "\">" + sd.present() + "</b></td>");
         for (String s : sorted(refs)) {
           b.append(genOutcome(items, s, si, sp));
         }
@@ -1790,31 +1606,31 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       }
     }
     if (ok)
-      return "<td style=\"background-color: #e6ffe6\"><a href=\""+item.getId()+".html\">\u2714</a></td>";
+      return "<td style=\"background-color: #e6ffe6\"><a href=\"" + item.getId() + ".html\">\u2714</a></td>";
     else
-      return "<td style=\"background-color: #ffe6e6\"><a href=\""+item.getId()+".html\">\u2716</a></td>";
+      return "<td style=\"background-color: #ffe6e6\"><a href=\"" + item.getId() + ".html\">\u2716</a></td>";
   }
 
   public void unzip(String zipFilePath, String destDirectory) throws IOException {
     File destDir = new File(destDirectory);
     if (!destDir.exists()) {
-        destDir.mkdir();
+      destDir.mkdir();
     }
     ZipInputStream zipIn = new ZipInputStream(new FileInputStream(zipFilePath));
     ZipEntry entry = zipIn.getNextEntry();
     // iterates over entries in the zip file
     while (entry != null) {
-        String filePath = destDirectory + File.separator + entry.getName();
-        if (!entry.isDirectory()) {
-            // if the entry is a file, extracts it
-            extractFile(zipIn, filePath);
-        } else {
-            // if the entry is a directory, make the directory
-            File dir = new File(filePath);
-            dir.mkdir();
-        }
-        zipIn.closeEntry();
-        entry = zipIn.getNextEntry();
+      String filePath = destDirectory + File.separator + entry.getName();
+      if (!entry.isDirectory()) {
+        // if the entry is a file, extracts it
+        extractFile(zipIn, filePath);
+      } else {
+        // if the entry is a directory, make the directory
+        File dir = new File(filePath);
+        dir.mkdir();
+      }
+      zipIn.closeEntry();
+      entry = zipIn.getNextEntry();
     }
     zipIn.close();
   }
@@ -1829,7 +1645,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       bos.write(bytesIn, 0, read);
     }
     bos.close();
-}
+  }
 
   private void download(String address, String filename) throws IOException {
     URL url = new URL(address);
@@ -1857,11 +1673,11 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     StringBuilder b = new StringBuilder();
     b.append("<html>");
     b.append("<head>");
-    b.append("<title>"+title+"</title>");
+    b.append("<title>" + title + "</title>");
     b.append("<link rel=\"stylesheet\" href=\"fhir.css\"/>\r\n");
     b.append("</head>");
     b.append("<body>");
-    b.append("<h2>"+title+"</h2>");
+    b.append("<h2>" + title + "</h2>");
     b.append(s);
     b.append("</body>");
     b.append("</html>");
@@ -1911,16 +1727,17 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     if (context.fetchResource(Resource.class, url) != null)
       return true;
     if (Utilities.existsInList(url, "http://hl7.org/fhir/sid/us-ssn", "http://hl7.org/fhir/sid/cvx", "http://hl7.org/fhir/sid/ndc", "http://hl7.org/fhir/sid/us-npi", "http://hl7.org/fhir/sid/icd-10",
-       "http://hl7.org/fhir/sid/icd-10-vn", "http://hl7.org/fhir/sid/icd-10-cm", "http://hl7.org/fhir/sid/icd-9-cm", "http://hl7.org/fhir/w5", "http://hl7.org/fhir/fivews",
-       "http://hl7.org/fhir/workflow", "http://hl7.org/fhir/ConsentPolicy/opt-out", "http://hl7.org/fhir/ConsentPolicy/opt-in")) {
+      "http://hl7.org/fhir/sid/icd-10-vn", "http://hl7.org/fhir/sid/icd-10-cm", "http://hl7.org/fhir/sid/icd-9-cm", "http://hl7.org/fhir/w5", "http://hl7.org/fhir/fivews",
+      "http://hl7.org/fhir/workflow", "http://hl7.org/fhir/ConsentPolicy/opt-out", "http://hl7.org/fhir/ConsentPolicy/opt-in")) {
       return true;
     }
     if (Utilities.existsInList(url, "http://loinc.org", "http://unitsofmeasure.org", "http://snomed.info/sct")) {
-       return true;
-     }
+      return true;
+    }
     if (fetcher != null) {
       return fetcher.resolveURL(appContext, path, url, type);
-    };
+    }
+    ;
     return false;
   }
 
@@ -1949,8 +1766,8 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       c.setDoOutput(true);
       c.setDoInput(true);
       c.setRequestMethod("POST");
-      c.setRequestProperty( "Content-type", "application/fhir+xml");
-      c.setRequestProperty( "Accept", "application/fhir+xml" );
+      c.setRequestProperty("Content-type", "application/fhir+xml");
+      c.setRequestProperty("Accept", "application/fhir+xml");
       c.getOutputStream().write(bs.toByteArray());
       c.getOutputStream().close();
 
@@ -1961,7 +1778,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
 //          System.out.println(line);
 //        }
 //        reader.close();
-        throw new IOException("Unable to PUT to "+output+": "+c.getResponseMessage());
+        throw new IOException("Unable to PUT to " + output + ": " + c.getResponseMessage());
       }
     } else {
       FileOutputStream s = new FileOutputStream(output);
@@ -1978,20 +1795,20 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         new org.hl7.fhir.dstu3.formats.XmlParser().setOutputStyle(org.hl7.fhir.dstu3.formats.IParser.OutputStyle.PRETTY).compose(s, res);
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         new org.hl7.fhir.dstu3.formats.JsonParser().setOutputStyle(org.hl7.fhir.dstu3.formats.IParser.OutputStyle.PRETTY).compose(s, res);
-      else if (fn.endsWith(".txt") || fn.endsWith(".map") )
+      else if (fn.endsWith(".txt") || fn.endsWith(".map"))
         TextFile.stringToStream(org.hl7.fhir.dstu3.utils.StructureMapUtilities.render((org.hl7.fhir.dstu3.model.StructureMap) res), s, false);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
     } else if (version.startsWith("4.0")) {
       org.hl7.fhir.r4.model.Resource res = VersionConvertor_40_50.convertResource(r);
       if (fn.endsWith(".xml") && !fn.endsWith("template.xml"))
         new org.hl7.fhir.r4.formats.XmlParser().setOutputStyle(org.hl7.fhir.r4.formats.IParser.OutputStyle.PRETTY).compose(s, res);
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         new org.hl7.fhir.r4.formats.JsonParser().setOutputStyle(org.hl7.fhir.r4.formats.IParser.OutputStyle.PRETTY).compose(s, res);
-      else if (fn.endsWith(".txt") || fn.endsWith(".map") )
+      else if (fn.endsWith(".txt") || fn.endsWith(".map"))
         TextFile.stringToStream(org.hl7.fhir.r4.utils.StructureMapUtilities.render((org.hl7.fhir.r4.model.StructureMap) res), s, false);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
     } else if (version.startsWith("1.4")) {
       org.hl7.fhir.dstu2016may.model.Resource res = VersionConvertor_14_50.convertResource(r);
       if (fn.endsWith(".xml") && !fn.endsWith("template.xml"))
@@ -1999,7 +1816,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         new org.hl7.fhir.dstu2016may.formats.JsonParser().setOutputStyle(org.hl7.fhir.dstu2016may.formats.IParser.OutputStyle.PRETTY).compose(s, res);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
     } else if (version.startsWith("1.0")) {
       VersionConvertorAdvisor50 advisor = new org.hl7.fhir.convertors.misc.IGR2ConvertorAdvisor5();
       org.hl7.fhir.dstu2.model.Resource res = VersionConvertor_10_50.convertResource(r, advisor);
@@ -2008,63 +1825,20 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         new org.hl7.fhir.dstu2.formats.JsonParser().setOutputStyle(org.hl7.fhir.dstu2.formats.IParser.OutputStyle.PRETTY).compose(s, res);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
     } else if (version.equals(Constants.VERSION)) {
       if (fn.endsWith(".xml") && !fn.endsWith("template.xml"))
         new XmlParser().setOutputStyle(org.hl7.fhir.r5.formats.IParser.OutputStyle.PRETTY).compose(s, r);
       else if (fn.endsWith(".json") && !fn.endsWith("template.json"))
         new JsonParser().setOutputStyle(org.hl7.fhir.r5.formats.IParser.OutputStyle.PRETTY).compose(s, r);
-      else if (fn.endsWith(".txt") || fn.endsWith(".map") )
-        TextFile.stringToStream(org.hl7.fhir.r5.utils.StructureMapUtilities.render((org.hl7.fhir.r5.model.StructureMap) r), s, false);
+      else if (fn.endsWith(".txt") || fn.endsWith(".map"))
+        TextFile.stringToStream(StructureMapUtilities.render((org.hl7.fhir.r5.model.StructureMap) r), s, false);
       else
-        throw new FHIRException("Unsupported format for "+fn);
+        throw new FHIRException("Unsupported format for " + fn);
     } else
-      throw new FHIRException("Encounted unsupported configured version "+version+" loading "+fn);
+      throw new FHIRException("Encounted unsupported configured version " + version + " loading " + fn);
 
     s.close();
-  }
-
-  public void setSnomedExtension(String sct) {
-    context.getExpansionParameters().addParameter("system-version", "http://snomed.info/sct|http://snomed.info/sct/"+sct);
-  }
-
-  public IValidatorResourceFetcher getFetcher() {
-    return fetcher;
-  }
-
-  public void setFetcher(IValidatorResourceFetcher fetcher) {
-    this.fetcher = fetcher;
-  }
-
-  public void setAssumeValidRestReferences(boolean assumeValidRestReferences) {
-    this.assumeValidRestReferences = assumeValidRestReferences;
-  }
-
-  public boolean isNoExtensibleBindingMessages() {
-    return noExtensibleBindingMessages;
-  }
-
-  public void setNoExtensibleBindingMessages(boolean noExtensibleBindingMessages) {
-    this.noExtensibleBindingMessages = noExtensibleBindingMessages;
-  }
-
-
-
-  public boolean isSecurityChecks() {
-    return securityChecks;
-  }
-
-  public void setSecurityChecks(boolean securityChecks) {
-    this.securityChecks = securityChecks;
-  }
-
-
-  public boolean isCrumbTrails() {
-    return crumbTrails;
-  }
-
-  public void setCrumbTrails(boolean crumbTrails) {
-    this.crumbTrails = crumbTrails;
   }
 
   public byte[] transformVersion(String source, String targetVer, FhirFormat format, Boolean canDoNative) throws FHIRException, IOException, Exception {
@@ -2083,10 +1857,10 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
         } else if (VersionUtilities.isR4Ver(version)) {
           return convertVersionNativeR4(targetVer, cnt, format);
         } else {
-          throw new FHIRException("Source version not supported yet: "+version);
+          throw new FHIRException("Source version not supported yet: " + version);
         }
       } catch (Exception e) {
-        System.out.println("Conversion failed using Java convertor: "+e.getMessage());
+        System.out.println("Conversion failed using Java convertor: " + e.getMessage());
       }
     }
     // ok, we try converting using the structure maps
@@ -2098,7 +1872,7 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     StructureMapUtilities scu = new StructureMapUtilities(context, new TransformSupportServices(outputs));
     StructureMap map = context.getTransform(url);
     if (map == null)
-      throw new Error("Unable to find map "+url+" (Known Maps = "+context.listMapUrls()+")");
+      throw new Error("Unable to find map " + url + " (Known Maps = " + context.listMapUrls() + ")");
     org.hl7.fhir.r5.elementmodel.Element resource = getTargetResourceFromStructureMap(map);
     scu.transform(null, src, map, resource);
     ByteArrayOutputStream bs = new ByteArrayOutputStream();
@@ -2109,226 +1883,225 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   private String getMapId(String type, String targetVer) {
     if (VersionUtilities.isR2Ver(version)) {
       if (VersionUtilities.isR3Ver(targetVer)) {
-        return "http://hl7.org/fhir/StructureMap/"+type+"2to3";
+        return "http://hl7.org/fhir/StructureMap/" + type + "2to3";
       }
     } else if (VersionUtilities.isR3Ver(version)) {
       if (VersionUtilities.isR2Ver(targetVer)) {
-        return "http://hl7.org/fhir/StructureMap/"+type+"3to2";
+        return "http://hl7.org/fhir/StructureMap/" + type + "3to2";
       } else if (VersionUtilities.isR4Ver(targetVer)) {
-        return "http://hl7.org/fhir/StructureMap/"+type+"3to4";
+        return "http://hl7.org/fhir/StructureMap/" + type + "3to4";
       }
     } else if (VersionUtilities.isR4Ver(version)) {
       if (VersionUtilities.isR3Ver(targetVer)) {
-        return "http://hl7.org/fhir/StructureMap/"+type+"4to3";
+        return "http://hl7.org/fhir/StructureMap/" + type + "4to3";
       }
     }
-    throw new FHIRException("Source/Target version not supported: "+version+" -> "+targetVer);
+    throw new FHIRException("Source/Target version not supported: " + version + " -> " + targetVer);
   }
 
   public byte[] convertVersionNativeR2(String targetVer, Content cnt, FhirFormat format) throws IOException, Exception {
     org.hl7.fhir.dstu2.model.Resource r2;
     switch (cnt.cntType) {
-    case JSON:
-      r2 = new org.hl7.fhir.dstu2.formats.JsonParser().parse(cnt.focus);
-      break;
-    case XML:
-      r2 = new org.hl7.fhir.dstu2.formats.XmlParser().parse(cnt.focus);
-      break;
-    default:
-      throw new FHIRException("Unsupported input format: "+cnt.cntType.toString());
+      case JSON:
+        r2 = new org.hl7.fhir.dstu2.formats.JsonParser().parse(cnt.focus);
+        break;
+      case XML:
+        r2 = new org.hl7.fhir.dstu2.formats.XmlParser().parse(cnt.focus);
+        break;
+      default:
+        throw new FHIRException("Unsupported input format: " + cnt.cntType.toString());
     }
     if (VersionUtilities.isR2Ver(targetVer)) {
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR2BVer(targetVer)) {
       org.hl7.fhir.dstu3.model.Resource r3 = VersionConvertor_10_30.convertResource(r2);
       org.hl7.fhir.dstu2016may.model.Resource r2b = VersionConvertor_14_30.convertResource(r3);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR3Ver(targetVer)) {
       org.hl7.fhir.dstu3.model.Resource r3 = VersionConvertor_10_30.convertResource(r2);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR4Ver(targetVer)) {
       org.hl7.fhir.r4.model.Resource r4 = VersionConvertor_10_40.convertResource(r2);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else {
-      throw new FHIRException("Target Version not supported yet: "+targetVer);
+      throw new FHIRException("Target Version not supported yet: " + targetVer);
     }
   }
 
   public byte[] convertVersionNativeR2b(String targetVer, Content cnt, FhirFormat format) throws IOException, Exception {
     org.hl7.fhir.dstu2016may.model.Resource r2b;
     switch (cnt.cntType) {
-    case JSON:
-      r2b = new org.hl7.fhir.dstu2016may.formats.JsonParser().parse(cnt.focus);
-      break;
-    case XML:
-      r2b = new org.hl7.fhir.dstu2016may.formats.XmlParser().parse(cnt.focus);
-      break;
-    default:
-      throw new FHIRException("Unsupported input format: "+cnt.cntType.toString());
+      case JSON:
+        r2b = new org.hl7.fhir.dstu2016may.formats.JsonParser().parse(cnt.focus);
+        break;
+      case XML:
+        r2b = new org.hl7.fhir.dstu2016may.formats.XmlParser().parse(cnt.focus);
+        break;
+      default:
+        throw new FHIRException("Unsupported input format: " + cnt.cntType.toString());
     }
     if (VersionUtilities.isR2Ver(targetVer)) {
       org.hl7.fhir.dstu3.model.Resource r3 = VersionConvertor_14_30.convertResource(r2b);
       org.hl7.fhir.dstu2.model.Resource r2 = VersionConvertor_10_30.convertResource(r3);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR2BVer(targetVer)) {
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR3Ver(targetVer)) {
       org.hl7.fhir.dstu3.model.Resource r3 = VersionConvertor_14_30.convertResource(r2b);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR4Ver(targetVer)) {
       org.hl7.fhir.r4.model.Resource r4 = VersionConvertor_14_40.convertResource(r2b);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else {
-      throw new FHIRException("Target Version not supported yet: "+targetVer);
+      throw new FHIRException("Target Version not supported yet: " + targetVer);
     }
   }
-
 
   public byte[] convertVersionNativeR3(String targetVer, Content cnt, FhirFormat format) throws IOException, Exception {
     org.hl7.fhir.dstu3.model.Resource r3;
     switch (cnt.cntType) {
-    case JSON:
-      r3 = new org.hl7.fhir.dstu3.formats.JsonParser().parse(cnt.focus);
-      break;
-    case XML:
-      r3 = new org.hl7.fhir.dstu3.formats.XmlParser().parse(cnt.focus);
-      break;
-    default:
-      throw new FHIRException("Unsupported input format: "+cnt.cntType.toString());
+      case JSON:
+        r3 = new org.hl7.fhir.dstu3.formats.JsonParser().parse(cnt.focus);
+        break;
+      case XML:
+        r3 = new org.hl7.fhir.dstu3.formats.XmlParser().parse(cnt.focus);
+        break;
+      default:
+        throw new FHIRException("Unsupported input format: " + cnt.cntType.toString());
     }
     if (VersionUtilities.isR2Ver(targetVer)) {
       org.hl7.fhir.dstu2.model.Resource r2 = VersionConvertor_10_30.convertResource(r3);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR2BVer(targetVer)) {
       org.hl7.fhir.dstu2016may.model.Resource r2b = VersionConvertor_14_30.convertResource(r3);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR3Ver(targetVer)) {
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR4Ver(targetVer)) {
       org.hl7.fhir.r4.model.Resource r4 = VersionConvertor_30_40.convertResource(r3, false);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else {
-      throw new FHIRException("Target Version not supported yet: "+targetVer);
+      throw new FHIRException("Target Version not supported yet: " + targetVer);
     }
   }
 
@@ -2336,68 +2109,68 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
   public byte[] convertVersionNativeR4(String targetVer, Content cnt, FhirFormat format) throws IOException, Exception {
     org.hl7.fhir.r4.model.Resource r4;
     switch (cnt.cntType) {
-    case JSON:
-      r4 = new org.hl7.fhir.r4.formats.JsonParser().parse(cnt.focus);
-      break;
-    case XML:
-      r4 = new org.hl7.fhir.r4.formats.XmlParser().parse(cnt.focus);
-      break;
-    default:
-      throw new FHIRException("Unsupported input format: "+cnt.cntType.toString());
+      case JSON:
+        r4 = new org.hl7.fhir.r4.formats.JsonParser().parse(cnt.focus);
+        break;
+      case XML:
+        r4 = new org.hl7.fhir.r4.formats.XmlParser().parse(cnt.focus);
+        break;
+      default:
+        throw new FHIRException("Unsupported input format: " + cnt.cntType.toString());
     }
     if (VersionUtilities.isR2Ver(targetVer)) {
       org.hl7.fhir.dstu2.model.Resource r2 = VersionConvertor_10_40.convertResource(r4);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2.formats.JsonParser().compose(bs, r2);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2.formats.XmlParser().compose(bs, r2);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR2BVer(targetVer)) {
       org.hl7.fhir.dstu2016may.model.Resource r2b = VersionConvertor_14_40.convertResource(r4);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu2016may.formats.JsonParser().compose(bs, r2b);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu2016may.formats.XmlParser().compose(bs, r2b);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR3Ver(targetVer)) {
       org.hl7.fhir.dstu3.model.Resource r3 = VersionConvertor_30_40.convertResource(r4, false);
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.dstu3.formats.JsonParser().compose(bs, r3);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.dstu3.formats.XmlParser().compose(bs, r3);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else if (VersionUtilities.isR4Ver(targetVer)) {
       ByteArrayOutputStream bs = new ByteArrayOutputStream();
       switch (format) {
-      case JSON:
-        new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
-        return bs.toByteArray();
-      case XML:
-        new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
-        return bs.toByteArray();
-      default:
-        throw new FHIRException("Unsupported output format: "+cnt.cntType.toString());
+        case JSON:
+          new org.hl7.fhir.r4.formats.JsonParser().compose(bs, r4);
+          return bs.toByteArray();
+        case XML:
+          new org.hl7.fhir.r4.formats.XmlParser().compose(bs, r4);
+          return bs.toByteArray();
+        default:
+          throw new FHIRException("Unsupported output format: " + cnt.cntType.toString());
       }
     } else {
-      throw new FHIRException("Target Version not supported yet: "+targetVer);
+      throw new FHIRException("Target Version not supported yet: " + targetVer);
     }
   }
 
@@ -2408,39 +2181,194 @@ public class ValidationEngine implements IValidatorResourceFetcher, IPackageInst
     return TextFile.streamToBytes(c.getInputStream());
   }
 
-  public FilesystemPackageCacheManager getPcm() {
-    return pcm;
-  }
-
-  public List<BundleValidationRule> getBundleValidationRules() {
-    return bundleValidationRules;
-   }
-
   public boolean packageExists(String id, String ver) throws IOException, FHIRException {
     return pcm.packageExists(id, ver);
   }
 
   public void loadPackage(String id, String ver) throws IOException, FHIRException {
-    loadIg(id+(ver == null ? "" : "#"+ver), true);
+    loadIg(id + (ver == null ? "" : "#" + ver), true);
   }
 
   /**
    * Systems that host the ValidationEngine can use this to control what validation the validator performs.
-   *
+   * <p>
    * Using this, you can turn particular kinds of validation on and off. In addition, you can override
    * the error | warning | hint level and make it a different level.
-   *
+   * <p>
    * Each entry has
    * * 'allowed': a boolean flag. if this is false, the Validator will not report the error.
    * * 'level' : set to error, warning, information
-   *
+   * <p>
    * Entries are registered by ID, using the IDs in /org.hl7.fhir.utilities/src/main/resources/Messages.properties
-   *
+   * <p>
    * This feature is not supported by the validator CLI - and won't be. It's for systems hosting
    * the validation framework in their own implementation context
    */
   public Map<String, ValidationControl> getValidationControl() {
     return validationControl;
+  }
+
+  public String setTerminologyServer(String src, String log, FhirPublication version) throws FHIRException, URISyntaxException {
+    return connectToTSServer(src, log, version);
+  }
+
+  /*
+   * Accessors
+   */
+
+  public boolean isDoNative() {
+    return doNative;
+  }
+
+  public ValidationEngine setNative(boolean doNative) {
+    this.doNative = doNative;
+    return this;
+  }
+
+  public boolean isNoInvariantChecks() {
+    return noInvariantChecks;
+  }
+
+  public ValidationEngine setNoInvariantChecks(boolean value) {
+    this.noInvariantChecks = value;
+    return this;
+  }
+
+  public boolean isHintAboutNonMustSupport() {
+    return hintAboutNonMustSupport;
+  }
+
+  public ValidationEngine setHintAboutNonMustSupport(boolean hintAboutNonMustSupport) {
+    this.hintAboutNonMustSupport = hintAboutNonMustSupport;
+    return this;
+  }
+
+  public boolean isAnyExtensionsAllowed() {
+    return anyExtensionsAllowed;
+  }
+
+  public ValidationEngine setAnyExtensionsAllowed(boolean anyExtensionsAllowed) {
+    this.anyExtensionsAllowed = anyExtensionsAllowed;
+    return this;
+  }
+
+  public String getVersion() {
+    return version;
+  }
+
+  public ValidationEngine setVersion(String version) {
+    this.version = version;
+    return this;
+  }
+
+  public String getLanguage() {
+    return language;
+  }
+
+  public ValidationEngine setLanguage(String language) {
+    this.language = language;
+    return this;
+  }
+
+  public FilesystemPackageCacheManager getPcm() {
+    return pcm;
+  }
+
+  public PrintWriter getMapLog() {
+    return mapLog;
+  }
+
+  public ValidationEngine setMapLog(String mapLog) throws FileNotFoundException {
+    if (mapLog != null) {
+      this.mapLog = new PrintWriter(mapLog);
+    }
+    return this;
+  }
+
+  public boolean isDebug() {
+    return debug;
+  }
+
+  public ValidationEngine setDebug(boolean debug) {
+    this.debug = debug;
+    return this;
+  }
+
+  public IValidatorResourceFetcher getFetcher() {
+    return fetcher;
+  }
+
+  public ValidationEngine setFetcher(IValidatorResourceFetcher fetcher) {
+    this.fetcher = fetcher;
+    return this;
+  }
+
+  public boolean isAssumeValidRestReferences() {
+    return assumeValidRestReferences;
+  }
+
+  public ValidationEngine setAssumeValidRestReferences(boolean assumeValidRestReferences) {
+    this.assumeValidRestReferences = assumeValidRestReferences;
+    return this;
+  }
+
+  public boolean isNoExtensibleBindingMessages() {
+    return noExtensibleBindingMessages;
+  }
+
+  public ValidationEngine setNoExtensibleBindingMessages(boolean noExtensibleBindingMessages) {
+    this.noExtensibleBindingMessages = noExtensibleBindingMessages;
+    return this;
+  }
+
+  public boolean isSecurityChecks() {
+    return securityChecks;
+  }
+
+  public ValidationEngine setSecurityChecks(boolean securityChecks) {
+    this.securityChecks = securityChecks;
+    return this;
+  }
+
+  public boolean isCrumbTrails() {
+    return crumbTrails;
+  }
+
+  public ValidationEngine setCrumbTrails(boolean crumbTrails) {
+    this.crumbTrails = crumbTrails;
+    return this;
+  }
+
+  public boolean isShowTimes() {
+    return showTimes;
+  }
+
+  public ValidationEngine setShowTimes(boolean showTimes) {
+    this.showTimes = showTimes;
+    return this;
+  }
+
+  public List<BundleValidationRule> getBundleValidationRules() {
+    return bundleValidationRules;
+  }
+
+  public ValidationEngine setBundleValidationRules(List<BundleValidationRule> bundleValidationRules) {
+    this.bundleValidationRules = bundleValidationRules;
+    return this;
+  }
+
+  public QuestionnaireMode getQuestionnaireMode() {
+    return questionnaireMode;
+  }
+
+  public ValidationEngine setQuestionnaireMode(QuestionnaireMode questionnaireMode) {
+    this.questionnaireMode = questionnaireMode;
+    return this;
+  }
+
+  public ValidationEngine setSnomedExtension(String sct) {
+    context.getExpansionParameters().addParameter("system-version", "http://snomed.info/sct|http://snomed.info/sct/" + sct);
+    return this;
   }
 
 }
