+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/cli/services/HTMLOutputGenerator.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.hl7.fhir.utilities.Utilities;
 import org.hl7.fhir.utilities.validation.ValidationMessage;
 import org.hl7.fhir.utilities.validation.ValidationMessage.IssueSeverity;
-import org.hl7.fhir.validation.ValidationEngine.ValidationRecord;
+import org.hl7.fhir.validation.ValidationRecord;
 import org.hl7.fhir.validation.cli.utils.VersionUtil;
 
 public class HTMLOutputGenerator {
